% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BYUImachine.R
\docType{package}
\name{BYUImachine}
\alias{BYUImachine}
\alias{BYUImachine-package}
\title{BYUImachine: A package for producing a robust prediction model using machine
learning to predict attrition of eClaims customers.}
\description{
Description goes here
}
\section{Functions}{

\code{preprocess_data}

\code{other_function}
}

