% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_measures.R
\name{extract_measures}
\alias{extract_measures}
\title{Extract measures from list of models}
\usage{
extract_measures(models, test_x, test_y)
}
\arguments{
\item{models}{List of models of class \code{train}}

\item{test_x}{`data.frame` or `tibble`. explanitory variables}

\item{test_y}{`vector` target variable}
}
\value{
This function returns a \code{tibble} of measures by model
}
\description{
Generates tibble of measures
}
\details{
This is a function that gives a list of measures of how well certain models
did. It's used within the \code{get_performance}.
}
\seealso{
\link[caret]{confusionMatrix}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
