% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_predictions_help_file.R
\name{get_likely_drops}
\alias{get_likely_drops}
\title{Find at risk customers}
\usage{
get_likely_drops(x, test_data, threshold)
}
\arguments{
\item{x}{A list of models}

\item{test_data}{a df, the portion of data you are using to test your predictions.}

\item{threshold}{a value between 0 and 1. What ratio of columns do you need to agree on the "Dropped" value}
}
\value{
This function returns a \code{tibble}, a single column reporting the ratio of like predictions for dropped clients meeting the prediction threshold
}
\description{
Makes predictions using a list of models and returns a vector where each value represents a customer which has been predicted to drop by a pre-specified percentage of models.
}
\author{
"Chad Schaeffer <sch12059@byui.edu>"
}
