% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance.R
\name{get_performance}
\alias{get_performance}
\title{Extract model performance}
\usage{
get_performance(list_of_models, pred = test_x, target = test_y)
}
\arguments{
\item{list_of_models}{List of models of class \code{train}}

\item{pred}{data.frame or tibble of predictor variables}

\item{target}{vector of target variable}
}
\value{
This function returns a \code{tibble} of model performance including columns:
\itemize{
 \item model
 \item measure
 \item score
}
}
\description{
Generates tibble of performance measures by model, measure and arranged by score
}
\details{
This is the details section
}
\seealso{
\link[BYUImachine]{extract_measures}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
