% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_var_imp.R
\name{gg_var_imp}
\alias{gg_var_imp}
\title{Visualize extracted overall feature importance}
\usage{
gg_var_imp(data, top_num = 20)
}
\arguments{
\item{data}{tibble of results from \code{HSPSModelR::var_imp_overall()}}

\item{top_num}{\code{integer} number of top features to display}
}
\value{
\code{tibble} of ranked features including columns:
\itemize{
 \item model
 \item feature
 \item Overall - Scaled score produced by \code{caret::varImp()}
 \item rank - Ordered rank within each model
}
}
\description{
Generates ggplot of top ranked features
}
\details{
To be used after data is generated from \code{HSPSModelR::var_imp_overall()}
}
\examples{
\dontrun{

vars <- var_imp_overall(models_list)

gg_var_imp(vars, top_num = 20)

}
}
\seealso{
\link[caret]{varImp}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
