% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_train_packages.R
\name{install_train_packages}
\alias{install_train_packages}
\title{Get packages needed for ensemble learning}
\usage{
install_train_packages()
}
\description{
Install and / or load all libraries used in the BYUImachine::run_models function
}
\examples{

\dontrun{
install_train_packages()
}
}
\author{
"Chad Schaeffer <sch12059@byui.edu>"
}
