% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Make a machine learning diagnostics table}
\usage{
make_table(models, test_x, test_y)
}
\arguments{
\item{models}{A list of models of class `train`}

\item{test_x}{`data.frame` or `tibble`. explanitory variables}

\item{test_y}{`vector` target variable}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item Type
 \item Accuracy - Overall performance of model
 \item Accuracy Lower
 \item Accuracy Null
 \item Accuracy P value
 \item Accuracy Upper
 \item Balanced Accuracy
 \item Detection Prevalence
 \item Detection Rate
 \item F1 - Hybrid metric usefull fr unbalanced classes
 \item Kappa - Compares an observed accuracy with an expected accuracy
 \item McNemar P Value
 \item Negagive Prediction Value
 \item Positive Predictive Value
 \item Precision - How accurate the positive predictions are
 \item Prevalence
 \item Recall - True positive rate, number of instances from the positive class
   that actually predictoed correctly
 \item Sensitivity - Same as recall
 \item Specificty - Number of instances from the negative class that were
   actually predicted correctly
 \item Method the algorithm used to train each particular model
}
}
\description{
Takes a list of trained machine learning models and returns diagnostics as a
  data frame as to compare the effectiveness of algorithms. Measures include
  Accuracy, Prevalence, Detection Rate, F1, Cohen's Kappa, McNemar P-Value,
  Negative and Positive Predictive value, Precision, Recall, Sensitivity, and
  Specificity
}
\examples{

\dontrun{
make_table(models_list, test_x, test_y)
}
}
\author{
"Chad Schaeffer <sch12059@byui.edu>
}
