% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Pre-process cognostic data}
\usage{
numeric_summary(x, na.rm = FALSE)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{y}{Class column}

\item{corr.cutoff}{Correlation coeff to exclude variables having correlation
above but not to exceed 1}

\item{freqCut}{[Insert def from caret]}

\item{uniqueCut}{[Insert def from caret]}

\item{impute.method}{Impute by "knn" or "mean"}
}
\value{
This function returns a \code{tibble} of optimized features
}
\description{
Generates a tibble with features optimized for machine learning
}
\details{
This is the details section
}
\examples{
data_summary(iris)
data_summary(airquality, na.rm = FALSE)

}
\seealso{
\link[base]{summary}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
