% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Pre-process cognostic data}
\usage{
preprocess_data(x, target = "Truth", reduce_cols = FALSE,
  factor_y = TRUE, impute = "zero", corr_cutoff = 0.9,
  freq_cut = 95/5, unique_cut = 10, k = 10)
}
\arguments{
\item{x}{data frame or tibble.}

\item{target}{classifier column}

\item{reduce_cols}{lgl `TRUE`: Columns are reduced based on near zero variance and correlation; FALSE = Nothing}

\item{factor_y}{`FALSE`: Recodes pred to 0 and 1; TRUE = Recodes pred to factor}

\item{impute}{character Impute NA by "knn","mean","zero"}

\item{corr_cutoff}{Corelation coefficient level to cut off highly correlated columns, devaulted to .90}

\item{freq_cut}{the cutoff for the ratio of the most common value to the second most common value}

\item{unique_cut}{the cutoff for the percentage of distinct values out of the number of total samples
(knn takes substantially longer to compute, zero replaces NA with 0)}

\item{k}{the number of nearest neighbours to use for imputate (defaults to 10)}
}
\value{
This function returns a \code{tibble} of optimized features
}
\description{
Generates a tibble with features optimized for machine learning
}
\details{
This function will .... reguardless of inputs
}
\seealso{
\link[caret]{preProcess}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
