% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp_overall.R
\name{var_imp_overall}
\alias{var_imp_overall}
\title{Extract overall feature importance within and between models}
\usage{
var_imp_overall(models)
}
\arguments{
\item{models}{List of models of class \code{train}}
}
\value{
\code{tibble} of ranked features including columns:
\itemize{
 \item model
 \item feature
 \item Overall scaled score produced by \code{caret::varImp()}
 \item rank ordered rank within each model
}
}
\description{
Generates tibble of ranked features based on within model importance and
between model agreeableness
}
\details{
To be used like \code{HSPSModelR::var_imp_raw()} but goes a step further
and estimates importance for each distinct feature
}
\examples{

\dontrun{

var_imp_overall(models_list)

}
}
\seealso{
\link[caret]{varImp}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
