% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp_raw.R
\name{var_imp_raw}
\alias{var_imp_raw}
\title{Extract feature importance within models}
\usage{
var_imp_raw(models)
}
\arguments{
\item{models}{List of models of class \code{train}}
}
\value{
This function returns a \code{tibble} of ranked features by model including columns:
\itemize{
 \item model
 \item feature
 \item Overall scaled score produced by \code{caret::varImp()}
 \item rank ordered rank within each model
}
}
\description{
Generates tibble of features ranked by importance by each model
}
\details{
This is the details section
}
\examples{

\dontrun{

var_imp_raw(models_list)

}
}
\seealso{
\link[caret]{varImp}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
