% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BYUImachine.R
\docType{package}
\name{BYUImachine}
\alias{BYUImachine}
\alias{BYUImachine-package}
\title{BYUImachine: A package for producing a robust prediction model using machine
learning to predict attrition of eClaims customers.}
\description{
THis package was made for Henry Schein, but it's meant to make the caret functions simpler and easier, and to build out a list of machine learning models that can either be used in an ensemble, or just use the one with the best predictors.
}
\section{Functions}{

\code{preprocess_data}

\code{gg_var_imp}

\code{var_imp_overall}

\code{var_imp_raw}

\code{install_train_packages}

\code{extract_measures}

\code{get_common_predictions}

\code{get_performance}

\code{make_list}

\code{make_table}

\code{run_models}
}

