% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_measures.R
\name{extract_measures}
\alias{extract_measures}
\title{Extract measures from list of models}
\usage{
extract_measures(models)
}
\arguments{
\item{models}{List of models of class \code{train}}
}
\value{
This function returns a \code{tibble} of measures by model
}
\description{
Generates tibble of measures
}
\details{
This is the details section
}
\seealso{
\link[caret]{confusionMatrix}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
