% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance.R
\name{get_performance}
\alias{get_performance}
\title{Extract model performance}
\usage{
get_performance(models, test_x, test_y)
}
\arguments{
\item{models}{List of models of class \code{train}}

\item{test_x}{`data.frame` or `tibble` of explanitory variables}

\item{test_y}{vector of target variable}
}
\value{
This function returns a \code{tibble} of model performance including columns:
\itemize{
 \item model
 \item measure
 \item score
}
}
\description{
Generates tibble of performance measures by model, measure and arranged by score
}
\details{
This is the details section
}
\seealso{
\link[BYUImachine]{extract_measures}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
