% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance.R
\name{get_performance}
\alias{get_performance}
\title{Extract model performance}
\usage{
get_performance(models, test_x, test_y, format = "long")
}
\arguments{
\item{models}{List of models of class \code{train}}

\item{test_x}{`data.frame` or `tibble` of explanitory variables}

\item{test_y}{vector of target variable}

\item{format}{"long" for long format (default), or "wide" for wide format}
}
\value{
This function returns a \code{tibble} of model performance including columns:
\itemize{
 \item model model name
 \item measure machine learning metrics
 \item score double, usually a range between 0 and 1
}
When \code{format = "wide"}, the function will return a \code{tibble} including columns:
\itemize{
 \item Type
 \item Accuracy - Overall performance of model
 \item Accuracy Lower
 \item Accuracy Null
 \item Accuracy P value
 \item Accuracy Upper
 \item Balanced Accuracy
 \item Detection Prevalence
 \item Detection Rate
 \item F1 - Hybrid metric usefull fr unbalanced classes
 \item Kappa - Compares an observed accuracy with an expected accuracy
 \item McNemar P Value
 \item Negagive Prediction Value
 \item Positive Predictive Value
 \item Precision - How accurate the positive predictions are
 \item Prevalence
 \item Recall - True positive rate, number of instances from the positive class
   that actually predictoed correctly
 \item Sensitivity - Same as recall
 \item Specificty - Number of instances from the negative class that were
   actually predicted correctly
 \item Method the algorithm used to train each particular model
}
}
\description{
Generates tibble of performance measures by model, measure and arranged by
  score
}
\examples{
\dontrun{
# Long format
p <- get_performance(models_list, test_x, test_y)
p
p \%>\% filter(measure == "F1")

# Wide format
get_performance(models_list, test_x, test_y, format = "wide")
}
}
\seealso{
\link[BYUImachine]{extract_measures}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
