% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_list.R
\name{make_list}
\alias{make_list}
\title{Make a list of models}
\usage{
make_list(x)
}
\arguments{
\item{x}{source or file path}
}
\value{
This function returns a \code{list} of contents at specified location
}
\description{
Compiles the contents of a folder into a list. Intended for machine learning
  models, and for use with the HSPSModelR::make_table function
}
\examples{

\dontrun{
models <- make_list("path/to/file")
}
}
\author{
"Chad Schaeffer <sch12059@byui.edu>, Brad Borget <bor13001@byui.edu>"
}
