% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_help_file.R
\name{make_table}
\alias{make_table}
\title{Make a machine learning diagnostics table}
\usage{
make_table(x, test_data, predict_var)
}
\arguments{
\item{x}{A list of models}

\item{test_data}{portion of data you are using to test your predictions.}

\item{predict_var}{the true values which you are comparing to your predicted values.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item Type
 \item Accuracy
 \item Accuracy Lower
 \item Accuracy Null
 \item Accuracy P value
 \item Accuracy Upper
 \item Balanced Accuracy
 \item Detection Prevalence
 \item Detection Rate
 \item F1
 \item Kappa
 \item McNemar P Value
 \item Negagive Prediction Value
 \item Positive Predictive Value
 \item Precision
 \item Prevalence
 \item Recall
 \item Sensitivity
 \item Specificty
 \item Method
}
}
\description{
Takes a list of trained machine learning models and returns diagnostics as a data frame, as to compare the effectiveness of algorithms. Measures include Accuracy, Prevalence, Detection Rate, F1, Cohen's Kappa, McNemar P-Value, Negative and Positive Predictive value, Precision, Recall, Sensitivity, and Specificity
}
\author{
"Chad Schaeffer <sch12059@byui.edu>
}
