% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Pre-process cognostic data}
\usage{
preprocess_data(x, pred = "Truth", corr_cutoff = 0.9,
  freq_cut = 95/5, unique_cut = 10, impute = "zero", k = 10,
  binary_y = FALSE, pca = FALSE)
}
\arguments{
\item{x}{data frame or tibble.}

\item{pred}{classifier column}

\item{k}{the number of nearest neighbours to use for imputate (defaults to 10)}

\item{binary_y}{TRUE = Recodes pred to 0 and 1; FALSE = Recodes pred to factor}

\item{pca}{TRUE = PCA is computed instead of corr}

\item{corr.cutoff}{A numeric value for the pair-wise absolute correlation cutoff (defaults to .90)}

\item{freqCut}{the cutoff for the ratio of the most common value to the second most common value}

\item{uniqueCut}{the cutoff for the percentage of distinct values out of the number of total samples}

\item{impute.method}{Impute NA by "knn","mean","zero", or "pca"
(knn takes substantially longer to compute, zero replaces NA with 0)}
}
\value{
This function returns a \code{tibble} of optimized features
}
\description{
Generates a tibble with features optimized for machine learning
}
\details{
This is the details section
}
\examples{
insert examples here

}
\seealso{
\link[base]{summary}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
