% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_models}
\alias{run_models}
\title{Run many models on train data}
\usage{
run_models(train_x, train_y, seed = 1, num_folds = 2,
  trim_models = TRUE, light = FALSE, prepro_methods = NULL)
}
\arguments{
\item{train_x}{data.frame or tibble of predictor variables}

\item{train_y}{vector of target variable}

\item{seed}{Integer. Seed for caretEnsemble::caretList()}

\item{num_folds}{Integer. Number of folds.}

\item{trim_models}{Logical. If TRUE, excess will be omited from models}

\item{light}{`logical.` FALSE data will train on only 5 algorithms. if TRUE
(default), data will train on all 68 models}

\item{prepro_methods}{prepro_methods string or vector of strings of
preprocessing methods. See also \link[caret]{preProcess}}
}
\value{
This function returns a \code{list} of trained models:
}
\description{
Generates list of models of class train
}
\examples{

\dontrun{

models_list <- run_models(train_x     = train_x,
                          train_y     = train_y,
                          seed        = 1,
                          num_folds   = 2,
                          trim_models = TRUE,
                          light       = TRUE)


}
}
\seealso{
\link[caretEnsemble]{caretList}
}
\author{
"Dallin Webb <dallinwebb@byui.edu>"
}
